requirements.txt
requests
re
lxml.html
selenium
gspread
datetime
chromedriver-binary==102.0.5005.61.0